+++ Warehouse.php
@@ -15,7 +15,7 @@
  * @package RosarioSIS
  */
 
-define( 'ROSARIO_VERSION', '2.9.5' );
+define( 'ROSARIO_VERSION', '2.9.6' );
 
 /**
  * Include config.inc.php file.