+++ CHANGES.md
@@ -8,6 +8,7 @@ Changes in 2.9.6
 - Fix SQL error invalid input syntax for type date in Discipline/*Breakdown.php
 - Fix #159 Update Help text to drop/remove a course in Help_en.php & Help_es.php
 - Fix #161 Letter Grade Widget search terms when combined in Widgets.fnc.php
+- Fix PHP7 error 'continue' not in the 'loop' or 'switch' context
 
 Changes in 2.9.5
 ----------------